Fix keyboard on macOS
I previously used command as meta, and then swapped that with option
because of how terminal apps works. And then a virtual machine Linux
also worked that way, so I thought that was the canonical way. Until I
noticed that VMWare swapped the keys for virtual machines, and on
native Linux and Windows alt is meta, and that's where the command is
on macOS. So swapping yet again, and it's more ergonomic for me
anyways.