Switch from moe-theme to tomorrow
moe-light is a bit too yellow and saturated. Solarized would have been
a better choice for the yellowish theme. But let's try a full blown
white one. Some of the color choices are a bit weird, but I'll
customize them bit by bit.

For example whitespace errors are too gray, and to me the statusbar
colors seem inverted (active/unactive).