Support for bytecode compiled init
It's not really working in exactly the same way as non-compiled. Some
references seem to be missing, some packages doesn't work the same
way, and I'm not initializing package.el which means that package
commands are not working until they are manually initialized.

Probably not worth the hassle, as it only saves half a second from
startup. But I'm just trying things out and learning.