Configure yasnippet a bit further
Load it with a mode hook instead of deferring it. More reliable and
doesn't stop text input at random points for a long period of time.

Also reload the snippets with a hook, which might make the error I get
when compiling emacs.org go away.

And fix smartparens conflict and allow sub-snippets.