Defer magit for speed but enable with a hook
Demanding magit made Emacs startup take almost double the time on my
computer. But I had to demand it for magit-wip to be enabled.
Otherwise I could code quite a long time before magit was loaded as I
didn't check stuff in git.

Now I have a find-file-hook that required magit if the file is under
git. Makes the start fast, makes the first git file load slow. And I
also load it after an idle delay, we'll see if that is unnecessary or
slow.