Roughly remove org tangling
It's more convenient for me to write elisp in a proper elisp buffer
rather than always popping open org-mode source edit buffer. When I
removed org I noticed a lot of typos and errors because flycheck
noticed them immediately. Also a lot of the configuration was all
around the place, which I didn't really notice as I usually used the
org file with everything folded.

As a bonus feature removing org also sped up loading by 25% or so.

Also did quite a lot of cleanup. But more should and will be done.
Some of the cleanup was stylistical, the other was removing disabled
packages and removing Emacs < 26 support.