Support Microsoft Python Language Server
Enabled with eglot if vj-mspy-dotnet-path and vj-mspy-path has been
defined. It's hard to get a proper commit of mspy that actually works,
at the moment commit: 681ccd5 seems to work ok. But I'm still in
test-mode with this.

Palantir Python Language Server is ok, but it bugs out way too often
and just stops completing.